# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /call-logs:
    get:
      operationId: listCallLogs
      tags: [calllogs]
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: agentId
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/CallStatus" }
        - name: dispositionId
          in: query
          required: false
          schema: { type: string }
        - name: direction
          in: query
          required: false
          schema: { $ref: "#/components/schemas/CallDir" }
        - name: hasRecording
          in: query
          required: false
          schema: { type: boolean }
        - name: dateFrom
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: dateTo
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: page
          in: query
          required: false
          schema: { type: integer }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLogList"
  /call-logs/{id}:
    get:
      operationId: getCallLog
      tags: [calllogs]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
  /call-logs/{id}/disposition:
    patch:
      operationId: updateCallLogDisposition
      tags: [calllogs]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispositionAssignInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
  /call-logs/{id}/retry:
    post:
      operationId: retryCall
      tags: [calllogs]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerCallResult"
  /dispositions:
    get:
      operationId: listDispositions
      tags: [dispositions]
      parameters:
        - name: agentId
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disposition"
    post:
      operationId: createDisposition
      tags: [dispositions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispositionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disposition"
  /dispositions/{id}:
    patch:
      operationId: updateDisposition
      tags: [dispositions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispositionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disposition"
    delete:
      operationId: deleteDisposition
      tags: [dispositions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    patch:
      operationId: updateAutomation
      tags: [automations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
  /agents/test-call:
    post:
      operationId: testCall
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCallInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerCallResult"
...
  /phone-numbers/{id}/agent:
    post:
      operationId: setInboundAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundAgentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationResult"
    delete:
      operationId: removeInboundAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationResult"
...
components:
  schemas:
...
    CallDir:
      type: string
      enum: [OUTBOUND, INBOUND]
    CallStatus:
      type: string
      enum: [INITIATED, RINGING, IN_PROGRESS, COMPLETED, FAILED, NO_ANSWER, BUSY, CANCELLED]
    AutomationType:
      type: string
      enum: [AUTO_CALL_ON_LEAD, RETRY_ON_DROP, SCHEDULED_FOLLOWUP, MONTHLY_CHECKIN]
...
    TriggerCallResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        call_log_id: { type: ["string", "null"] }
        execution_id: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
...
    CallLog:
      type: object
      required: [id, org_id, bolna_execution_id, bolna_agent_id, direction, phone_number, status, started_at, drop_detected, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        lead_id: { type: ["string", "null"] }
        lead_name: { type: ["string", "null"] }
        bolna_execution_id: { type: string }
        bolna_agent_id: { type: string }
        agent_name: { type: ["string", "null"] }
        direction: { $ref: "#/components/schemas/CallDir" }
        phone_number: { type: string }
        status: { $ref: "#/components/schemas/CallStatus" }
        started_at: { type: string, format: date-time }
        ended_at: { type: ["string", "null"], format: date-time }
        duration_seconds: { type: ["integer", "null"] }
        transcript: { type: ["string", "null"] }
        summary: { type: ["string", "null"] }
        recording_url: { type: ["string", "null"] }
        disposition_id: { type: ["string", "null"] }
        disposition: { oneOf: [ { $ref: "#/components/schemas/Disposition" }, { type: "null" } ] }
        drop_detected: { type: boolean }
        drop_reason: { type: ["string", "null"] }
        retry_of_call_id: { type: ["string", "null"] }
        retry_call_id: { type: ["string", "null"] }
        call_type: { type: ["string", "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CallLogList:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/CallLog" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    DispositionAssignInput:
      type: object
      required: [disposition_id]
      properties:
        disposition_id: { type: string }
    Disposition:
      type: object
      required: [id, org_id, bolna_agent_id, label, color, is_active, sort_order, created_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        bolna_agent_id: { type: string }
        label: { type: string }
        color: { type: string }
        description: { type: ["string", "null"] }
        is_active: { type: boolean }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
    DispositionInput:
      type: object
      required: [bolna_agent_id, label]
      properties:
        bolna_agent_id: { type: string }
        label: { type: string, minLength: 1 }
        color: { type: string }
        description: { type: string }
        sort_order: { type: integer }
    DispositionUpdate:
      type: object
      properties:
        label: { type: string }
        color: { type: string }
        description: { type: ["string", "null"] }
        is_active: { type: boolean }
        sort_order: { type: integer }
    Automation:
      type: object
      required: [id, org_id, name, type, bolna_agent_id, trigger_config, is_active, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        name: { type: string }
        type: { $ref: "#/components/schemas/AutomationType" }
        bolna_agent_id: { type: string }
        trigger_config: { type: object, additionalProperties: true }
        is_active: { type: boolean }
        last_triggered_at: { type: ["string", "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AutomationInput:
      type: object
      required: [name, type, bolna_agent_id]
      properties:
        name: { type: string, minLength: 1 }
        type: { $ref: "#/components/schemas/AutomationType" }
        bolna_agent_id: { type: string }
        trigger_config: { type: object, additionalProperties: true }
        is_active: { type: boolean }
    AutomationUpdate:
      type: object
      properties:
        name: { type: string }
        bolna_agent_id: { type: string }
        trigger_config: { type: object, additionalProperties: true }
        is_active: { type: boolean }
...
    TestCallInput:
      type: object
      required: [agent_id, phone]
      properties:
        agent_id: { type: string }
        phone: { type: string }
        name: { type: string }
    InboundAgentInput:
      type: object
      required: [agent_id]
      properties:
        agent_id: { type: string }
    MutationResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        error: { type: ["string", "null"] }
...